+++ package.json
@@ -17,19 +17,28 @@
   "files": [
     "dist"
   ],
-  "main": "./dist/index.js",
+  "main": "./dist/index.cjs",
   "module": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "exports": {
-    ".": "./dist/index.js",
+    ".": {
+      "import": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/index.js"
+      },
+      "require": {
+        "types": "./dist/index.d.cts",
+        "default": "./dist/index.cjs"
+      }
+    },
     "./package.json": "./package.json"
   },
   "publishConfig": {
     "access": "public"
   },
   "scripts": {
-    "build": "tsdown",
-    "dev": "tsdown --watch",
+    "build": "rollup -c rollup.config.mjs",
+    "dev": "rollup -c -w",
     "test": "vitest",
     "typecheck": "tsc --noEmit",
     "lint": "eslint --config eslint.config.ts",
@@ -37,12 +46,15 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.34.0",
+    "@rollup/plugin-typescript": "^12.1.4",
     "@types/node": "^22.15.17",
     "bumpp": "^10.1.0",
     "eslint": "^9.34.0",
     "globals": "^16.3.0",
     "jiti": "^2.5.1",
-    "tsdown": "^0.11.9",
+    "rollup": "^4.48.1",
+    "rollup-plugin-dts": "^6.2.3",
+    "tslib": "^2.8.1",
     "typescript": "^5.8.3",
     "typescript-eslint": "^8.40.0",
     "vitest": "^3.1.3"
