+++ rollup.config.mjs
@@ -1,0 +1,148 @@
+import fs from "node:fs";
+// import { createRequire } from "node:module";
+import { dirname, join } from "node:path";
+import { fileURLToPath } from "node:url";
+import dtsPlugin from "rollup-plugin-dts";
+import tsPlugin from "@rollup/plugin-typescript";
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+
+const testPatterns = ["**/*.bench.ts", "**/*.spec.ts", "**/*.test.ts"];
+
+// const packageJson = createRequire(import.meta.url)("./package.json");
+
+export default () => {
+  clearDir("dist");
+
+  /**
+   * TODO
+   * package.json의 exports를 entrypoints로 사용하고
+   * publishConfig의 export를 번들링에 사용하려고 했는데
+   * 문제가 해결되지 않아 entrypoints를 하드코딩함
+   */
+  const entrypoints = ["src/index.ts"];
+  // const entrypoints = Object.values(packageJson.exports).filter(
+  //   (f) => /^(\.\/)?src\//.test(f) && f.endsWith(".ts")
+  // );
+
+  return [
+    libBuildOptions({
+      format: "esm",
+      extension: "js",
+      entrypoints,
+      outDir: "dist",
+      sourcemap: false,
+    }),
+    libBuildOptions({
+      format: "cjs",
+      extension: "cjs",
+      entrypoints,
+      outDir: "dist",
+      sourcemap: false,
+    }),
+    declarationOptions({
+      entrypoints,
+      outDir: "dist",
+    }),
+  ];
+};
+
+/**
+ * @type {(options: {
+ *   entrypoints: string[];
+ *   format: 'esm' | 'cjs';
+ *   extension: 'js' | 'cjs' | 'mjs';
+ *   outDir: string;
+ *   sourcemap: boolean;
+ * }) => import('rollup').RollupOptions}
+ */
+function libBuildOptions({
+  entrypoints,
+  extension,
+  format,
+  outDir,
+  sourcemap,
+}) {
+  return {
+    input: mapInputs(entrypoints),
+    plugins: [
+      tsPlugin({
+        exclude: [...testPatterns],
+        compilerOptions: {
+          sourceMap: sourcemap,
+          inlineSources: sourcemap || undefined,
+          removeComments: !sourcemap,
+          declaration: false,
+        },
+      }),
+    ],
+    output: {
+      format,
+      dir: outDir,
+      ...fileNames(extension),
+      // Using preserveModules disables bundling and the creation of chunks,
+      // leading to a result that is a mirror of the input module graph.
+      preserveModules: true,
+      sourcemap,
+      generatedCode: "es2015",
+      // Hoisting transitive imports adds bare imports in modules,
+      // which can make imports by JS runtimes slightly faster,
+      // but makes the generated code harder to follow.
+      hoistTransitiveImports: false,
+    },
+  };
+}
+
+/**
+ * @type {(options: {entrypoints: string[]; outDir: string}) => import('rollup').RollupOptions}
+ */
+function declarationOptions({ entrypoints, outDir }) {
+  return {
+    plugins: [dtsPlugin()],
+    input: mapInputs(entrypoints),
+    output: [
+      {
+        format: "esm",
+        dir: outDir,
+        generatedCode: "es2015",
+        ...fileNames("d.ts"),
+        preserveModules: true,
+        preserveModulesRoot: "src",
+      },
+      {
+        format: "cjs",
+        dir: outDir,
+        generatedCode: "es2015",
+        ...fileNames("d.cts"),
+        preserveModules: true,
+        preserveModulesRoot: "src",
+      },
+    ],
+  };
+}
+
+/** @type {(srcFiles: string[]) => Record<string, string>} */
+function mapInputs(srcFiles) {
+  return Object.fromEntries(
+    srcFiles.map((file) => [
+      file.replace(/^(\.\/)?src\//, "").replace(/\.[cm]?(js|ts)$/, ""),
+      join(__dirname, file),
+    ])
+  );
+}
+
+function fileNames(extension = "js") {
+  return {
+    entryFileNames: `[name].${extension}`,
+    chunkFileNames: `_chunk/[name]-[hash:6].${extension}`,
+  };
+}
+
+/** @type {(dir: string) => void} */
+function clearDir(dir) {
+  const dirPath = join(__dirname, dir);
+  if (dir && fs.existsSync(dirPath)) {
+    fs.rmSync(dirPath, { recursive: true, force: true });
+    console.log(`cleared: ${dir}`);
+  }
+}
