+++ src/index.ts
@@ -1,3 +1,4 @@
-export * from "./core/business";
-export * from "./core/holiday";
-export * from "./types";
+export * from "./core/business.ts";
+export * from "./core/holiday.ts";
+export * from "./core/trading.ts";
+export * from "./types.ts";
