--- tsdown.config.ts
@@ -1,9 +1,0 @@
-import { defineConfig } from 'tsdown'
-
-export default defineConfig([
-  {
-    entry: ['./src/index.ts'],
-    platform: 'neutral',
-    dts: true,
-  },
-])
